requirements.txt
# Core dependencies
openai-whisper>=20231117

# OpenAI API client
openai>=1.3.0

# Optional: Alternative API services
assemblyai>=0.17.0

# Audio processing dependencies (usually installed with whisper)
torch>=1.9.0
torchaudio>=0.9.0
ffmpeg-python>=0.2.0

# Utility dependencies
requests>=2.25.0
pathlib2>=2.3.0; python_version < '3.4'

# Development dependencies (optional)
pytest>=6.0.0
black>=21.0.0
flake8>=3.8.0
mypy>=0.910
